+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.1-wip
+
+- Introduce `Builder` and `Linker` interface.
+- Copy `resources.json` to the build directory.
+
 ## 0.6.0
 
 - Add support for `hook/link.dart`.